// src/api/models/PlayerSynergy.ts
export interface PlayerSynergy {
    /**
     * 
     * @type {string}
     * @memberof PlayerSynergy
     */
    playerA: string;
    /**
     * 
     * @type {string}
     * @memberof PlayerSynergy
     */
    playerB: string;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    synergy: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    winProbDelta: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    gamesTogether: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    winsTogether: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    expectedWins: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    stdError: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    zScore: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    gamesApart: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    mainA: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    mainB: number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSynergy
     */
    adjustedExpectedWins: number;
}